requirements/base.txt
Django==4.2.19
djangorestframework==3.15.2
jsonmerge==1.9.2
drf-yasg==1.21.8
